rcar_gen3: Fix calculation method of VI6_UDS_HPHASE.HSTP/HEDP.
Changed the calculation method of VI6_UDS_HPHASE.HSTP/HEDP to the method
described in chapter 32.3.7.5 of the hardware manual.
This change improves the phenomenon that only the right edge extends when
the image is enlarged with UDS.

Since VSPM has its own margin processing, adjust with the value of
VI6_WPFn_HSZCLIP.HCL_OFST.

However, combining UDS and SRU may distort the image, so this patch is
effective only when using only UDS.
The conventional original calculation method of VSPM drv is applied to
the case where UDS and SRU are used simultaneously.

Signed-off-by: Koji <koji.matsuoka.xm@renesas.com>